multiple updates and improvements
updated the SOLID principles and Design Patterns projects with
additional notes and comments

added a new interview prep folder with common topics that come up in a
entry level developer role

the usual documentation improvements
