+++ DesignPatterns/DesignPatterns.sln
