+++ DesignPatterns/DesignPatterns/DesignPatterns.csproj
