+++ DesignPatterns/DesignPatterns/Program.cs
