+++ DesignPatterns/DesignPatterns/Properties/launchSettings.json
