+++ DesignPatterns/DesignPatterns/WeatherForecast.cs
