+++ DesignPatterns/DesignPatterns/appsettings.Development.json
