+++ InterviewPrep/index.md
@@ -1,0 +1,16 @@
+# Index of InterviewPrep Folder
+
+Below are links to each file in this folder:
+
+- [ADONET.md](ADONET.md)
+- [AgileMethodology.md](AgileMethodology.md)
+- [Angular.md](Angular.md)
+- [Blazor.md](Blazor.md)
+- [CSharpDotNetGeneral.md](CSharpDotNetGeneral.md)
+- [DesignPatterns.md](DesignPatterns.md)
+- [EntityFramework.md](EntityFramework.md)
+- [MSSQL.md](MSSQL.md)
+- [Razor.md](Razor.md)
+- [React.md](React.md)
+- [readme.md](readme.md)
+- [SOLID.md](SOLID.md)
