+++ InterviewPrep/readme.md
@@ -1,0 +1,33 @@
+# Interview Prep
+
+General Topics that come up in a standard .NET developer evaluation.
+
+- [ADONET.md](ADONET.md)
+- [AgileMethodology.md](AgileMethodology.md)
+- [Angular.md](Angular.md)
+- [Blazor.md](Blazor.md)
+- [CSharpDotNetGeneral.md](CSharpDotNetGeneral.md)
+- [DesignPatterns.md](DesignPatterns.md)
+- [EntityFramework.md](EntityFramework.md)
+- [MSSQL.md](MSSQL.md)
+- [Razor.md](Razor.md)
+- [React.md](React.md)
+- [readme.md](readme.md)
+- [SOLID.md](SOLID.md)
+
+# book a session with me
+
+1. [calendly](https://calendly.com/jaycodingtutor/30min)
+
+# hire and get to know me
+
+find ways to hire me, follow me and stay in touch with me.
+
+1. [github](https://github.com/Jay-study-nildana)
+1. [personal site](https://thechalakas.com)
+1. [upwork](https://www.upwork.com/fl/vijayasimhabr)
+1. [fiverr](https://www.fiverr.com/jay_codeguy)
+1. [codementor](https://www.codementor.io/@vijayasimhabr)
+1. [stackoverflow](https://stackoverflow.com/users/5338888/jay)
+1. [Jay's Coding Channel on YouTube](https://www.youtube.com/channel/UCJJVulg4J7POMdX0veuacXw/)
+1. [medium blog](https://medium.com/@vijayasimhabr)
