+++ SOLIDPrinciples/readme.md
@@ -15,8 +15,19 @@ Note 1 : The harrymt files were created in a older version of visual studio I th
 1. https://www.harrymt.com/blog/2018/05/16/SOLID-principles.html
 1. https://en.wikipedia.org/wiki/SOLID
 
+# book a session with me
+
+1. [calendly](https://calendly.com/jaycodingtutor/30min)
+
 # hire and get to know me
 
 find ways to hire me, follow me and stay in touch with me.
 
-https://jay-study-nildana.github.io/developerprofile/
+1. [github](https://github.com/Jay-study-nildana)
+1. [personal site](https://thechalakas.com)
+1. [upwork](https://www.upwork.com/fl/vijayasimhabr)
+1. [fiverr](https://www.fiverr.com/jay_codeguy)
+1. [codementor](https://www.codementor.io/@vijayasimhabr)
+1. [stackoverflow](https://stackoverflow.com/users/5338888/jay)
+1. [Jay's Coding Channel on YouTube](https://www.youtube.com/channel/UCJJVulg4J7POMdX0veuacXw/)
+1. [medium blog](https://medium.com/@vijayasimhabr)
